PR corrections
