+++ CHANGELOG.md
@@ -4,14 +4,12 @@ This changelog track changes to the qoqo project starting at version v0.5.0
 
 ## Unreleased
 
-* Added ControlledRotateX, ControlledRotateXY to qoqo
 * Added Three-qubit gates support
-* Added ControlledControlledPauliZ, ControlledControlledPhaseShift to qoqo
+* Added ControlledRotateX, ControlledRotateXY, ControlledControlledPauliZ, ControlledControlledPhaseShift to qoqo
 
 ## 1.2.5
 
 * Updated to pyo3 0.18.1
-
 
 ## 1.2.4
